ocr-app/backend/requirements.txt
fastapi
uvicorn[standard]
pillow
pytesseract
pdf2image
python-multipart
